+++ Studio/Analysis/ShapeScalarPanel.h
@@ -40,7 +40,7 @@ class ShapeScalarPanel : public QWidget {
 
   void run_clicked();
 
-  void handle_job_progress(int progress);
+  void handle_job_progress(double progress);
   void handle_job_complete();
 
  Q_SIGNALS: